hnatt/requirements.txt
absl-py==0.7.0
astor==0.6.2
bleach==1.5.0
click==6.7
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.8.2
Flask==1.0.2
gast==0.2.2
grpcio==1.13.0
h5py==2.8.0
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
Markdown==2.6.11
MarkupSafe==1.0
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
numpy==1.16.0
pandas==0.23.1
pathlib==1.0.1
plac==0.9.6
preshed==1.0.0
protobuf==3.6.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.12
regex==2017.4.5
scikit-learn==0.20.4
scipy==1.1.0
six==1.11.0
spacy==2.0.11
tensorboard==1.15.0
tensorflow-gpu==1.15.0
termcolor==1.1.0
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.4
ujson==2.0.0
Werkzeug==0.14.1
wrapt==1.11.1
